feat(25-01): alembic migration + ORM models for perp schema
- Migration 013: int->float for quantity/fill_quantity/shares columns
- Migration 013: add side column (default 'long') to orders and portfolio_holdings
- OrderRecord: quantity Mapped[float], added side field
- OrderFillRecord: fill_quantity Mapped[float]
- PortfolioHoldingRecord: shares Mapped[float], added side field

+++ src/poseidon/models/order_fill.py
@@ -20,7 +20,7 @@ class OrderFillRecord(Base):
         UUID(as_uuid=True), ForeignKey("orders.id"), nullable=False
     )
     fill_price: Mapped[float] = mapped_column(Float, nullable=False)
-    fill_quantity: Mapped[int] = mapped_column(Integer, nullable=False)
+    fill_quantity: Mapped[float] = mapped_column(Float, nullable=False)
     fill_time: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), nullable=False
     )
